[GUI] Refactor code
