+++ common/src/main/java/org/astraea/common/Utils.java
@@ -16,10 +16,14 @@
  */
 package org.astraea.common;
 
+import java.io.PrintWriter;
+import java.io.StringWriter;
 import java.net.InetAddress;
+import java.text.SimpleDateFormat;
 import java.time.Duration;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 import java.util.Set;
@@ -286,6 +290,21 @@ public final class Utils {
         .filter(obj -> obj instanceof String)
         .map(obj -> (String) obj)
         .collect(Collectors.toCollection(TreeSet::new));
+  }
+
+  public static String toString(Throwable e) {
+    var sw = new StringWriter();
+    var pw = new PrintWriter(sw);
+    e.printStackTrace(pw);
+    return sw.toString();
+  }
+
+  public static String format(long timestamp) {
+    if (timestamp > 0) {
+      var format = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss.SSS");
+      return format.format(new Date(timestamp));
+    }
+    return "unknown";
   }
 
   private Utils() {}
