+++ gui/src/main/java/org/astraea/gui/box/IntegerBox.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.gui;
+package org.astraea.gui.box;
 
 import java.util.Collection;
 import javafx.application.Platform;
