+++ gui/src/main/java/org/astraea/gui/tab/TopicTab.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.gui;
+package org.astraea.gui.tab;
 
 import java.util.List;
 import java.util.Map;
@@ -23,17 +23,20 @@ import java.util.stream.Collectors;
 import javafx.scene.control.Tab;
 import org.astraea.common.DataSize;
 import org.astraea.common.LinkedHashMap;
+import org.astraea.common.Utils;
 import org.astraea.common.admin.Broker;
 import org.astraea.common.admin.NodeInfo;
 import org.astraea.common.admin.Partition;
+import org.astraea.gui.Context;
+import org.astraea.gui.pane.PaneBuilder;
 
 public class TopicTab {
   public static Tab of(Context context) {
     var pane =
         PaneBuilder.of()
             .searchField("topic name")
-            .buttonTableAction(
-                input ->
+            .buttonAction(
+                (input, logger) ->
                     context.submit(
                         admin ->
                             admin
@@ -41,7 +44,7 @@ public class TopicTab {
                                 .thenApply(
                                     names ->
                                         names.stream()
-                                            .filter(name -> input.matchSearch(name))
+                                            .filter(input::matchSearch)
                                             .collect(Collectors.toSet()))
                                 .thenCompose(
                                     names ->
