+++ gui/src/main/java/org/astraea/gui/tab/UpdateTopicTab.java
@@ -14,63 +14,56 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.gui;
+package org.astraea.gui.tab;
 
 import java.util.HashMap;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import javafx.scene.control.Tab;
 import org.astraea.common.admin.TopicConfigs;
+import org.astraea.gui.Context;
+import org.astraea.gui.pane.PaneBuilder;
 
-public class CreateTopicTab {
+public class UpdateTopicTab {
 
   private static final String TOPIC_NAME = "topic";
   private static final String NUMBER_OF_PARTITIONS = "number of partitions";
-  private static final String NUMBER_OF_REPLICAS = "number of replicas";
 
   public static Tab of(Context context) {
 
     var pane =
         PaneBuilder.of()
-            .buttonName("CREATE")
+            .buttonName("UPDATE")
             .input(TOPIC_NAME, true, false)
             .input(NUMBER_OF_PARTITIONS, false, true)
-            .input(NUMBER_OF_REPLICAS, false, true)
-            .input(TopicConfigs.ALL_CONFIGS)
-            .buttonMessageAction(
-                input -> {
+            .input(TopicConfigs.DYNAMICAL_CONFIGS)
+            .buttonListener(
+                (input, logger) -> {
                   var allConfigs = new HashMap<>(input.texts());
                   var name = allConfigs.remove(TOPIC_NAME);
+                  var partitions = allConfigs.remove(NUMBER_OF_PARTITIONS);
                   return context.submit(
                       admin ->
                           admin
                               .topicNames(true)
                               .thenCompose(
                                   names -> {
-                                    if (names.contains(name))
-                                      return CompletableFuture.completedFuture(
-                                          name + " is already existent");
+                                    if (!names.contains(name))
+                                      return CompletableFuture.failedFuture(
+                                          new IllegalArgumentException(name + " is nonexistent"));
 
-                                    return admin
-                                        .creator()
-                                        .topic(name)
-                                        .numberOfPartitions(
-                                            Optional.ofNullable(
-                                                    allConfigs.remove(NUMBER_OF_PARTITIONS))
-                                                .map(Integer::parseInt)
-                                                .orElse(1))
-                                        .numberOfReplicas(
-                                            Optional.ofNullable(
-                                                    allConfigs.remove(NUMBER_OF_REPLICAS))
-                                                .map(Short::parseShort)
-                                                .orElse((short) 1))
-                                        .configs(allConfigs)
-                                        .run()
-                                        .thenApply(i -> "succeed to create topic:" + name);
+                                    return Optional.ofNullable(partitions)
+                                        .map(Integer::parseInt)
+                                        .map(ps -> admin.addPartitions(name, ps))
+                                        .orElse(CompletableFuture.completedFuture(null))
+                                        .thenCompose(
+                                            ignored -> admin.updateConfig(name, allConfigs))
+                                        .thenAccept(
+                                            ignored -> logger.log("succeed to update " + name));
                                   }));
                 })
             .build();
-    var tab = new Tab("create topic");
+    var tab = new Tab("update topic");
     tab.setContent(pane);
     return tab;
   }
