+++ gui/src/main/java/org/astraea/gui/text/TextField.java
@@ -1,0 +1,46 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.astraea.gui.text;
+
+public class TextField extends javafx.scene.control.TextField {
+  public static TextField onlyNumber() {
+    var field = new TextField();
+    field
+        .textProperty()
+        .addListener(
+            (observable, oldValue, newValue) -> {
+              if (!newValue.matches("\\d*")) {
+                field.setText(newValue.replaceAll("[^\\d]", ""));
+              }
+            });
+    return field;
+  }
+
+  public static TextField copyableField(String content) {
+    var field = new TextField(content);
+    field.setEditable(false);
+    return field;
+  }
+
+  public TextField() {
+    super();
+  }
+
+  public TextField(String text) {
+    super(text);
+  }
+}
